refactor(core): split fixflow.go and give the report state its own type
fixflow.go was 770 lines and was not "the fix flow". It held seven
unrelated concerns that shared a file because they all hung off Engine:
preview, apply, the post-apply re-check, batch, rollback, a checkpoint DTO
adapter, and the mutable report state.

It is now seven files named for what they hold — preview.go, modeplan.go,
apply.go, recheck.go, batch.go, rollback.go, checkpoints.go — with the
largest at 342 lines. Every function moved whole; the split is mechanical
and the line-by-line comparison below shows nothing else changed.

Two things did change, and both were the point of doing it.

The report state is its own type. Five Engine fields — mu, current, hasRun,
lastDelta, and the methods over them — existed only to serve it, and every
one of those methods touched nothing else: no registry, no store, no host.
Sharing a body with the rest of the engine bought nothing and cost the
invariant that matters, which is the copy in Current(). Report is a struct
but Findings is a slice, so returning one by value hands the caller the
same backing array markFixed writes into; that shipped once, and the race
detector caught it as a write in markFixed against json.Encoder serving
/api/result. While the state was five loose fields, copying was a rule to
remember at each read. Behind reportState there is one way out and it
copies. A test now asks that directly, in both directions, instead of
waiting for two goroutines to interleave.

verifyEdit is runEditValidator. It sat 100 lines from verifyFix, sharing a
prefix and nothing else: one runs a fix's declared validator against staged
bytes before the write, the other re-runs a checker after the apply to see
whether the finding is gone. FixOutcome.Verified and Action.VerifyCmd are
unrelated facts and read as though they were not. The two comments in
internal/fix and the sentence in AGENTS.md that named it are updated.

Behaviour is unchanged. Comparing the old file against the concatenated new
ones, ignoring package and import headers and normalising the rename, every
line of preview, apply, batch, rollback and checkpoints matches exactly;
the only differences are the report-state methods moving from an Engine
receiver to a reportState one.
