+++ AGENTS.md
@@ -119,7 +119,7 @@ Flow: `cmd/hostveil/app.go` builds the one engine (all eleven checkers + `fix.De
 - **A score with no axis behind it is not a perfect score.** `ScoreBreakdown.Applicable` is false when every domain was skipped or failed, and all three UIs render N/A instead of a number. The per-axis flag already stopped a skipped domain scoring 100; without this the same lie survived in the aggregate, since renormalizing over zero caps fell out at a clean 100 for a host nobody could look at.
 - **A score you can't improve by doing everything right measures nothing.** `RemediationUnavailable` findings are weighted down (`unavailableRelief`) because every image ships CVEs with no upstream patch; charging those in full pins the axis at 0 for a perfectly maintained host. They are not free either — the risk is real, and zero would be its own lie.
 - **A checker must never let "I couldn't look" pass for "nothing there."** The two score identically and mean opposite things, which is how a non-root scan once reported a perfect CVE score. So: `Available()` probes the *capability*, not just the binary (`platform.DockerReachable`, not `Has(r, "docker")`); a checker that covered only part of its ground returns its findings plus a `check.PartialError` (→ `ScanDegraded`); one that covered none returns an ordinary error (→ `ScanError`, axis excluded). Degraded axes *are* scored, flagged via `ScoreAxis.Degraded`, and every UI must render that flag — a total failure dressed as Degraded would restore the false 100 by another route.
-- **An edit fix may declare a validator, and it runs before anything is written.** `fix.Action.VerifyCmd` is argv containing `fix.VerifyPathToken`; `Engine.verifyEdit` stages the *original* and the *transformed* bytes in a temp dir and runs the command on each. Only a validator that accepts the original is trusted to reject the replacement — `sshd -t` needs to read the host keys, so where it cannot it fails on every config including the one in service, and without that control run the checker's own inability would block a correct fix. Checking before the write means a rejected fix leaves no file to restore and no checkpoint to explain. Every SSH fix carries `sshd -t`, because sshd keeps serving from the config it already loaded: a broken file looks like nothing at all until the next restart, and repairing it then needs the access it just removed.
+- **An edit fix may declare a validator, and it runs before anything is written.** `fix.Action.VerifyCmd` is argv containing `fix.VerifyPathToken`; `Engine.runEditValidator` stages the *original* and the *transformed* bytes in a temp dir and runs the command on each. Only a validator that accepts the original is trusted to reject the replacement — `sshd -t` needs to read the host keys, so where it cannot it fails on every config including the one in service, and without that control run the checker's own inability would block a correct fix. Checking before the write means a rejected fix leaves no file to restore and no checkpoint to explain. Every SSH fix carries `sshd -t`, because sshd keeps serving from the config it already loaded: a broken file looks like nothing at all until the next restart, and repairing it then needs the access it just removed.
 - **Apply order is always backup → write → checkpoint.** `applyEdit` refuses to write if the backup fails. Exec fixes have no rollback checkpoint (nothing file-backed), only a history record. The write itself goes through `writeFileAtomic` — temp file beside the target, then rename — because `os.WriteFile` truncates first, and a crash between truncate and write leaves a zero-length `sshd_config` whose backup you now need SSH to reach. `preserveOwner` carries uid/gid across the rename; hostveil runs as root, so skipping it would hand the operator's own compose file to root.
 - **One fix at a time, and a scan is a fix for this purpose.** `Engine.applyMu` serializes apply, batch, rollback, and scan. `applyEdit` is a read-modify-write and the dashboard serves concurrently, so without it two fixes to one compose file each read the original and the later write erased the earlier one — while both checkpoints recorded success and both findings were marked Fixed. `ApplyFix` takes the lock and `applyFix` is the unlocked body, because `ApplyBatch` loops and `sync.Mutex` is not reentrant.
 - **`Engine.Current()` returns a snapshot, not a view.** `Report` is a struct but `Findings` is a slice, so returning it by value shared the backing array that `markFixed` writes into — `go test -race` caught it as a write in `markFixed` against `json.Encoder` in the dashboard's `/api/result`. It clones.
